Migrated to 1.19.4

+++ src/main/java/gollorum/signpost/utils/math/geometry/Vector3.java
@@ -38,7 +38,7 @@ public final class Vector3 {
     }
 
     public BlockPos toBlockPos() {
-        return new BlockPos(x, y, z);
+        return new BlockPos((int)x, (int)y, (int)z);
     }
 
     public static Vector3 min(Vector3 a, Vector3 b) {
